feat: move annotation-processor to be under platform-spigot aggregator
