+++ .idea/encodings.xml
@@ -23,6 +23,8 @@
     <file url="file://$PROJECT_DIR$/modules/placeholder/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/modules/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/modules/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/platform-spigot/annotation-processor/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/platform-spigot/annotation-processor/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/platform-spigot/config-spigot/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/platform-spigot/config-spigot/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/platform-spigot/core-spigot/src/main/java" charset="UTF-8" />
