--- annotation-processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,0 @@
-tech.guilhermekaua.spigotboot.annotationprocessor.plugin.PluginAnnotationProcessor
