+++ platform-spigot/annotation-processor/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>tech.guilhermekaua.spigot-boot</groupId>
-        <artifactId>spigot-boot</artifactId>
+        <artifactId>spigot-boot-platform-spigot</artifactId>
         <version>2.0.2</version>
     </parent>
