+++ platform-spigot/annotation-processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,0 +1,1 @@
+tech.guilhermekaua.spigotboot.spigot.annotationprocessor.plugin.PluginAnnotationProcessor
