+++ pom.xml
@@ -43,7 +43,6 @@
         <module>utils</module>
         <module>config</module>
         <module>test-plugin</module>
-        <module>annotation-processor</module>
     </modules>
 
     <properties>
