+++ test-plugin/src/main/java/tech/guilhermekaua/spigotboot/testPlugin/Main.java
@@ -24,7 +24,6 @@ package tech.guilhermekaua.spigotboot.testPlugin;
 
 import lombok.Getter;
 import org.bukkit.plugin.java.JavaPlugin;
-import tech.guilhermekaua.spigotboot.annotationprocessor.annotations.Plugin;
 import tech.guilhermekaua.spigotboot.config.spigot.SpigotConfigModule;
 import tech.guilhermekaua.spigotboot.core.SpigotBoot;
 import tech.guilhermekaua.spigotboot.core.context.Context;
@@ -32,6 +31,7 @@ import tech.guilhermekaua.spigotboot.core.spigot.SpigotBootPlugin;
 import tech.guilhermekaua.spigotboot.core.spigot.SpigotCoreModule;
 import tech.guilhermekaua.spigotboot.data.ormLite.DataOrmLiteModule;
 import tech.guilhermekaua.spigotboot.placeholder.PlaceholderModule;
+import tech.guilhermekaua.spigotboot.spigot.annotationprocessor.annotations.Plugin;
 import tech.guilhermekaua.spigotboot.testPlugin.configuration.MainConfig;
 
 @Getter
